requirements.txt
PyYAML==6.0
dulwich==0.20.35
pre-commit==2.19.0
antlr4-python3-runtime==4.10
elasticsearch==8.2.0
PyMySQL==1.0.2
iteration-utilities==0.11.0
Flask==2.1.2
Flask-RESTful==0.3.9
flasgger==0.9.5
pika==1.2.1
requests==2.28.0
